ajustes

+++ src/main/java/br/com/api/funcionarios/controller/FinanceiroController.java
@@ -4,7 +4,6 @@ import br.com.api.funcionarios.service.FinanceiroService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -18,7 +17,7 @@ public class FinanceiroController {
 
     // 1- Recebe uma lista de funcionários, mês e ano. E retorna o valor total pago (salário e benefício) a esses funcionários no mês.
     @GetMapping("/total-value-list")
-    public BigDecimal calcBeneficioAnual(@RequestParam(name = "mes") int mes,
+    public double calcBeneficioAnual(@RequestParam(name = "mes") int mes,
                                          @RequestParam(name = "ano") int ano,
                                          @RequestParam(name = "listaFunc") List<Long> funcionarioList){
 
@@ -27,7 +26,7 @@ public class FinanceiroController {
 
     // 2- Recebe uma lista de funcionários, mês e ano. E retorna o total pago somente em salários no mês.
     @GetMapping("/total-salary-list")
-    public BigDecimal totalSalario(@RequestParam(name = "mes") int mes,
+    public double totalSalario(@RequestParam(name = "mes") int mes,
                               @RequestParam(name = "ano") int ano,
                               @RequestParam(name = "listaFunc") List<Long> funcionarioList){
 
@@ -36,7 +35,7 @@ public class FinanceiroController {
 
     // 3- Recebe uma lista somente com os funcionários que recebem benefícios, mês e ano. E retorna o total pago em benefícios no mês.
     @GetMapping("/total-beneficio-list")
-    public BigDecimal totalBeneficio(@RequestParam(name = "mes") int mes,
+    public double totalBeneficio(@RequestParam(name = "mes") int mes,
                                      @RequestParam(name = "ano") int ano,
                                      @RequestParam(name = "listaFunc") List<Long> funcionarioList){
